updated install and uninstall scripts.

+++ uninstall.pl
@@ -12,24 +12,36 @@ if(not defined($shell)){
 if(not defined($home)){
 	die "Env variable HOME is not defined!\n";
 }
+open UID,"id -u |";
+my $uid = <UID>;
+chomp($uid);
+$uid = $uid + 0;
 
 if($shell eq "bash"){
-	check_and_remove("/etc/bash.bashrc");
-	check_and_remove("/etc/bashrc");
+	if($uid == 0){
+		check_and_remove("/etc/bash.bashrc");
+		check_and_remove("/etc/bashrc");
+	}
 	check_and_remove("$home/.bashrc");
 }
 else{
-	check_and_remove("/etc/csh.cshrc");
-	check_and_remove("/etc/cshrc");
+	if($uid == 0){
+		check_and_remove("/etc/csh.cshrc");
+		check_and_remove("/etc/cshrc");
+	}
 	check_and_remove("$home/.cshrc");
 }
 
-if(-e "/usr/bin/trsh.pl"){
-	system("rm /usr/bin/trsh.pl");
+
+if($uid == 0){
+	if(-e "/usr/bin/trsh.pl"){
+		system("rm /usr/bin/trsh.pl");
+	}
+	if(-e "/usr/share/man/man1/trsh.1.gz"){
+		system("rm /usr/share/man/man1/trsh.1.gz");
+	}
 }
-if(-e "/usr/share/man/man1/trsh.1.gz"){
-	system("rm /usr/share/man/man1/trsh.1.gz");
-}
+
 if(-e "$home/.trsh.pl"){
 	system("rm $home/.trsh.pl");
 }
@@ -47,12 +59,16 @@ if(-e "$home/.Trash"){
 		print "Trash is at $home/.Trash. Delete its contents at your convienence\n";
 	}
 }
+if($uid != 0){
+	print "Just performed an user uninstallation. \n";
+	print "If you needed a system wide uninstall, do it as root.\n";
+}
 
 sub check_and_remove{
 	my $file = shift;
 	if(-e $file){
 		open RC,"$file" or die "Could not open $file to read\n";
-		open NRC,"+>$file.new" or die "Could not create $file\n";
+		open NRC,"+>$file.new" or die "Could not create $file.new\n";
 		my @contents = ();
 		while(my $line = <RC>){
 			if($line =~ /TRSH/){
@@ -62,7 +78,7 @@ sub check_and_remove{
 		}
 		close(RC);
 		close(NRC);
-		system("cp $file.new $file");
+		system("mv $file.new $file");
 	}
 }
